+++ alembic/env.py
@@ -6,6 +6,7 @@ from sqlalchemy import pool
 from alembic import context
 
 from app.config import create_settings
+from app.database.models import Base
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
@@ -16,13 +17,13 @@ config = context.config
 if config.config_file_name is not None:
     fileConfig(config.config_file_name)
 
-config.set_main_option("sqlalchemy.url", create_settings().postgres_url)
+config.set_main_option("sqlalchemy.url", create_settings().postgres_url + "?async_fallback=True")
 
 # add your model's MetaData object here
 # for 'autogenerate' support
 # from myapp import mymodel
 # target_metadata = mymodel.Base.metadata
-target_metadata = models.Base.metadata
+target_metadata = Base.metadata
 
 # other values from the config, defined by the needs of env.py,
 # can be acquired: